requirements.txt
# Core trading libraries
alpaca-py==0.42.0

# Data analysis and manipulation
pandas==2.3.1
numpy==2.2.6
yfinance==0.2.36

# Web framework
Flask==3.1.1

# Visualization
plotly==6.2.0
matplotlib==3.10.3

# Technical analysis
TA-Lib==0.4.28

# Configuration and environment
python-dotenv==1.0.0

# HTTP requests
requests==2.32.4

# Database (optional)
SQLAlchemy==2.0.0

# Logging and monitoring
structlog==23.2.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Development tools
black==23.7.0
flake8==6.0.0
mypy==1.5.1

# Documentation
Sphinx==7.2.0

# Additional utilities
beautifulsoup4==4.13.4
lxml==4.9.3 